requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
redis==7.4.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
aiohttp==3.13.4
